app/requirements.txt
matplotlib
numpy
pandas
scikit-learn
tensorflow
fastapi
uvicorn
jinja2
Pillow
python-multipart